testing_models/requirements.txt
numpy==1.23.5
albumentations==1.1.0
PyQt5>=5.15.9
opencv-python>=4.8.0.76
onnxruntime>=1.15.1
scipy>=1.7.0
requests>=2.25.1
tqdm>=4.61.0
tensorflow>=2.6.0
keras>=2.6.0
dlib==19.22.1
insightface==0.7.3
facenet-pytorch>=2.5.2
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
tensorboard>=2.6.0 